fix: restore progressive delivery validation contracts
