+++ src/autoskillit/core/types/_type_constants.py
@@ -741,7 +741,9 @@ DATA_MANIFEST_SOURCE_TYPES: frozenset[str] = frozenset(
 RUN_PYTHON_PATH_LIKE_ARGS: frozenset[str] = frozenset(
     {"output_dir", "workspace", "diagnostics_log_dir", "investigation_path"}
 )
-RUN_PYTHON_SENTINEL_KEYS: frozenset[str] = frozenset({"callable", "timeout", "work_dir"})
+RUN_PYTHON_SENTINEL_KEYS: frozenset[str] = frozenset(
+    {"callable", "step_name", "timeout", "work_dir"}
+)
 
 SCOPE_DIRECTION_SOURCE_TYPES: frozenset[str] = frozenset(
     {
