+++ src/autoskillit/server/_recipe_initialization.py
@@ -539,10 +539,9 @@ def _render_completion_receipt(
 
 
 def _public_completion_credential(
-    snapshot: RecipeExecutionSnapshot,
+    credential: RecipeExecutionCredential,
     projection: FinalizedRecipeProjection,
 ) -> RecipeExecutionCredential:
-    credential = build_recipe_execution_credential(snapshot)
     if not projection.delivery_segments:
         return credential
     initial_steps = frozenset(projection.delivery_segments[0].ordered_step_names)
@@ -599,7 +598,7 @@ def build_completion_response(
                 artifact_generation=state.artifact_generation,
                 flow_generation=state.flow_generation,
                 credential=_public_completion_credential(
-                    state.installed_execution.snapshot,
+                    build_recipe_execution_credential(state.installed_execution.snapshot),
                     state.finalized_projection,
                 ),
             )
@@ -624,7 +623,7 @@ def build_completion_response(
             artifact_generation=state.artifact_generation,
             flow_generation=state.flow_generation,
             credential=_public_completion_credential(
-                state.staged_snapshot,
+                build_recipe_execution_credential(state.staged_snapshot),
                 state.finalized_projection,
             ),
         )
