+++ tests/server/test_recipe_initialization.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import hashlib
 import json
-from types import SimpleNamespace
-from typing import cast
 from unittest.mock import Mock
 
 import pytest
@@ -20,7 +18,7 @@ from autoskillit.core import (
     FinalizedRecipeSegment,
     RecipeArtifactGeneration,
     RecipeBindingProjection,
-    RecipeExecutionSnapshot,
+    RecipeExecutionCredential,
     RecipeFlowGeneration,
 )
 from autoskillit.pipeline import (
@@ -182,18 +180,15 @@ def test_segmented_completion_credential_is_scoped_to_initial_bodies() -> None:
             FinalizedRecipeSegment(name="S1", ordered_step_names=("future",)),
         ),
     )
-    snapshot = cast(
-        RecipeExecutionSnapshot,
-        SimpleNamespace(
-            execution_id="execution",
-            snapshot_digest=_hash("snapshot"),
-            template_digests={"initial": _hash("initial"), "future": _hash("future")},
-        ),
+    credential = RecipeExecutionCredential(
+        execution_id="execution",
+        snapshot_digest=_hash("snapshot"),
+        invocation_template_digests={"initial": _hash("initial"), "future": _hash("future")},
     )
 
-    credential = recipe_initialization._public_completion_credential(snapshot, projection)
+    public_credential = recipe_initialization._public_completion_credential(credential, projection)
 
-    assert credential.invocation_template_digests == {"initial": _hash("initial")}
+    assert public_credential.invocation_template_digests == {"initial": _hash("initial")}
 
 
 def test_install_rejects_initializing_recipe_without_completion_receipt(
